backend/requirements.txt
arrow==1.2.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
black==22.3.0
blessed==1.19.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-extra-fields==3.4.1
et-xmlfile==1.1.0
filelock==3.6.0
ftfy==6.1.1
idna==3.4
importlib-metadata==5.1.0
inquirer==2.9.2
Jinja2==3.1.1
jinja2-time==0.2.0
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numexpr==2.8.3
oauthlib==3.2.2
openpyxl==3.0.10
pandantic==0.7.0
pathspec==0.9.0
Pillow==9.3.0
pip==22.3.1
pipenv==2022.4.21
platformdirs==2.5.2
poyo==0.5.0
pyarrow==8.0.0
pycparser==2.21
Pygments==2.12.0
PyInquirer==1.0.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==6.1.1
python-ulid==1.1.0
python3-openid==3.2.0
pytz==2022.1
readchar==3.0.5
regex==2022.4.24
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.3.0
setuptools==62.1.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.13
virtualenv-clone==0.5.7
wcwidth==0.2.5
wheel==0.38.4
zipp==3.11.0
